requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.6
dataclassy==0.6.2
decorator==4.4.2
ecdsa==0.16.1
et-xmlfile==1.0.1
future==0.18.2
geojson==2.5.0
grpcio==1.36.0
grpcio-tools==1.36.0
idna==2.10
importlib-metadata==3.7.3
iniconfig==1.1.1
networkx==2.5
numpy==1.20.1
openpyxl==3.0.7
packaging==20.9
pandapower==2.6.0
pandas==1.2.3
pluggy==0.13.1
protobuf==3.14.0
py==1.10.0
pycparser==2.20
pycryptodome==3.3.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-jose-cryptodome==1.3.2
pytz==2021.1
requests==2.25.1
scipy==1.6.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
xlrd==2.0.1
XlsxWriter==1.3.7
zepben.evolve==0.23.0b9
zepben.protobuf==0.13.0b4
zipp==3.4.1